+++ src/views/components/index.ts
@@ -27,3 +27,4 @@ export {
 } from '@views/components/MetadataFilters/MetadataFilters.tsx'
 export { MainMenu } from '@views/components/Header'
 export { CardScore } from '@views/components/CardScore/CardScore.tsx'
+export { CardHeader } from '@views/components/CardHeader/CardHeader.tsx'
